+++ src/composables/atc.ts
@@ -22,16 +22,16 @@ export const useFacilitiesIds = () => {
 export function getControllerPositionColor(controller: VatsimShortenedController) {
     const ids = useFacilitiesIds();
 
-    if (controller.isATIS) {
-        return radarColors.warning600;
+    if (controller.isATIS || controller.facility === ids.ATIS) {
+        return radarColors.warning500;
     }
 
     if (controller.facility === ids.DEL) {
-        return radarColors.primary600;
+        return radarColors.primary700;
     }
 
     if (controller.facility === ids.TWR) {
-        return radarColors.error500;
+        return radarColors.error700;
     }
 
     if (controller.facility === ids.GND) {
@@ -39,11 +39,11 @@ export function getControllerPositionColor(controller: VatsimShortenedController
     }
 
     if (controller.facility === ids.APP) {
-        return radarColors.info500;
+        return radarColors.error300;
     }
 
     if (controller.facility === ids.CTR) {
-        return radarColors.primary400;
+        return radarColors.primary300;
     }
 
     return radarColors.darkgray800;